Add initial BL implementations and update IConfig interface
Introduced `AdminImplementation`, `CallImplementation`, and
`VolunteerImplementation` classes in the `BlImplementation`
namespace, implementing `IAdmin`, `ICall`, and `IVolunteer`
interfaces respectively. Methods in `CallImplementation` and
`VolunteerImplementation` are placeholders throwing
`NotImplementedException`.

Extended the `IConfig` interface in the `DalApi` namespace to
include a new `RiskTimeSpan` property of type `TimeSpan`.

Removed the `<ItemGroup>` entry for the `BlImplementation`
folder from the `BL.csproj` file.